requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.8
cryptography==43.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0


